Заменены абсолютные пути на относительные в конфигурационном файле. Рассортированы методы по доступу. В конфигурационном файле переменные теперь все с маленькой буквы.

+++ ConsolePrograms/Log/CustomRollingFileAppender.cs
@@ -10,41 +10,44 @@ namespace ConsolePrograms {
     public class CustomRollingFileAppender: RollingFileAppender {
         
         public int MaxFileInDirectories { get; set; }
-        public int MaxFileInArhive { get; set; }
-        public string LocationLog { get; set; }
-        public string LocationLogArchive { get; set; }
+        public int MaxFileInArchive { get; set; }
+        public string? LocationLog { get; set; }
+        public string? LocationLogArchive { get; set; }
         
         public override void ActivateOptions() {
             base.ActivateOptions();
 
             Console.WriteLine($"MaxFileInDirectories: {MaxFileInDirectories}");
-            Console.WriteLine($"MaxFileInArhiv: {MaxFileInArhive}");
+            Console.WriteLine($"MaxFileInArhiv: {MaxFileInArchive}");
             Console.WriteLine($"LocationLog: {LocationLog}");
             Console.WriteLine($"LocationLogArchive: {LocationLogArchive}");
         }
         
-
         protected override void AdjustFileBeforeAppend() {
             ArchiveLogs(LocationLog, LocationLogArchive);
             base.AdjustFileBeforeAppend();
         }
 
-        public void ArchiveLogs(string sourceDirectory, string zipPath) {
-            IEnumerable<string> files = Directory.EnumerateFiles(sourceDirectory, "*.log");
-            IEnumerable<string> logFiles = files.Where(file => {
-                DateTime fileLastWriteTime = System.IO.File.GetLastWriteTime(file);
-                DateTime currentDate = DateTime.Now.Date;
-                DateTime fileDate = fileLastWriteTime.Date;
-                return (currentDate - fileDate).Days >= MaxFileInDirectories;
-            });
-            
-            using (ZipArchive zip = ZipFile.Open(zipPath, ZipArchiveMode.Update)) {
-                foreach (string logFile in logFiles) {
-                    string entryName = Path.Combine("archive", Path.GetFileName(logFile));
-                    zip.CreateEntryFromFile(logFile, entryName, CompressionLevel.Optimal);
-                    System.IO.File.Delete(logFile);
-                }
-                LimitLogFilesInArchive(zip);
+        private void ArchiveLogs(string? sourceDirectory, string? zipPath) {
+            if (sourceDirectory != null) {
+                IEnumerable<string> files = Directory.EnumerateFiles(sourceDirectory, "*.log");
+                IEnumerable<string> logFiles = files.Where(file => {
+                    DateTime fileLastWriteTime = System.IO.File.GetLastWriteTime(file);
+                    DateTime currentDate = DateTime.Now.Date;
+                    DateTime fileDate = fileLastWriteTime.Date;
+                    return (currentDate - fileDate).Days >= MaxFileInDirectories;
+                });
+
+                if (zipPath != null)
+                    using (ZipArchive zip = ZipFile.Open(zipPath, ZipArchiveMode.Update)) {
+                        foreach (string logFile in logFiles) {
+                            string entryName = Path.Combine("archive", Path.GetFileName(logFile));
+                            zip.CreateEntryFromFile(logFile, entryName, CompressionLevel.Optimal);
+                            System.IO.File.Delete(logFile);
+                        }
+
+                        LimitLogFilesInArchive(zip);
+                    }
             }
         }
         
@@ -55,7 +58,7 @@ namespace ConsolePrograms {
                 .ToList();
             
 
-            while (logEntries.Count > MaxFileInArhive) {
+            while (logEntries.Count > MaxFileInArchive) {
                 ZipArchiveEntry entryToDelete = logEntries.First();
                 entryToDelete.Delete();
                 logEntries.RemoveAt(0);
